+++ changelog.md
@@ -1,3 +1,7 @@
+#### 0.4.25 - 7/11/2019 ####
+Quantum Manufacturing upgrade<br>
+Quantum Swarm cost lowered from 465k to 450k<br>
+
 #### 0.4.24 - 7/10/2019 ####
 Thermomechanics upgrade for alloy production<br>
 Ziggurats unlockable via ancients upgrade<br>